refactor: project structure: added store folder
