+++ src/app/users/me/layout.tsx
@@ -1,10 +1,8 @@
 'use client'
 
 import React from 'react'
-import Link from 'next/link'
 
 import { Header } from '~/components/header'
-import { useNavigation } from '~/components/navigation-context'
 
 export default function Layout({
   children
