+++ src/app/workspaces/layout.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import Link from 'next/link'
 
 import { Header } from '~/components/header'
-import { useNavigation } from '~/components/navigation-context'
+import { useNavigation } from '~/components/store/navigation-context'
 
 export default function Layout({
   children
